Repo sync

+++ daos/pgsql/Items.php
@@ -21,18 +21,22 @@ class Items extends \daos\mysql\Items {
     public function get($options = array()) {
         $params = array();
         $where = '';
-        
+        $order = 'DESC';
         // only starred
         if(isset($options['type']) && $options['type']=='starred')
             $where .= ' AND starred=true ';
             
         // only unread
-        else if(isset($options['type']) && $options['type']=='unread')
+        else if(isset($options['type']) && $options['type']=='unread'){
             $where .= ' AND unread=true ';
+            if(\F3::get('unread_order')=='asc'){
+                $order = 'ASC';
+            }
+        }
         
         // search
         if(isset($options['search']) && strlen($options['search'])>0) {
-            $search = str_replace(" ", "%", trim($options['search']));
+            $search = implode('%', \helpers\Search::splitTerms($options['search']));
             $params[':search'] = $params[':search2'] = $params[':search3'] = array("%".$search."%", \PDO::PARAM_STR);
             $where .= ' AND (items.title LIKE :search OR items.content LIKE :search2 OR sources.title LIKE :search3) ';
         }
@@ -46,6 +50,12 @@ class Items extends \daos\mysql\Items {
         elseif(isset($options['source']) && strlen($options['source'])>0) {
             $params[':source'] = array($options['source'], \PDO::PARAM_INT);
             $where .= " AND items.source=:source ";
+        }
+
+        // update time filter
+        if(isset($options['updatedsince']) && strlen($options['updatedsince'])>0) {
+            $params[':updatedsince'] = array($options['updatedsince'], \PDO::PARAM_STR);
+            $where .= " AND items.updatetime > :updatedsince ";
         }
         
         // set limit
@@ -61,10 +71,10 @@ class Items extends \daos\mysql\Items {
 
         // get items from database
         return \F3::get('db')->exec('SELECT 
-                    items.id, datetime, items.title AS title, content, unread, starred, source, thumbnail, icon, uid, link, sources.title as sourcetitle, sources.tags as tags
+                    items.id, datetime, items.title AS title, content, unread, starred, source, thumbnail, icon, uid, link, updatetime, sources.title as sourcetitle, sources.tags as tags
                    FROM items, sources 
                    WHERE items.source=sources.id '.$where.' 
-                   ORDER BY items.datetime DESC 
+                   ORDER BY items.datetime '.$order.' 
                    LIMIT ' . $options['items'] . ' OFFSET ' . $options['offset'], $params);
     }
     
@@ -198,13 +208,17 @@ class Items extends \daos\mysql\Items {
     }
 
     /**
-     * cleanup old items
+     * cleanup orphaned and old items
      *
      * @return void
-     * @param DateTime $date date to delete all items older than this value
+     * @param DateTime $date date to delete all items older than this value [optional]
      */
-    public function cleanup(\DateTime $date) {
-        \F3::get('db')->exec('DELETE FROM items WHERE starred=false AND datetime<:date',
+    public function cleanup(\DateTime $date = NULL) {
+        \F3::get('db')->exec('DELETE FROM items WHERE id IN (
+                                SELECT items.id FROM items LEFT JOIN sources
+                                ON items.source=sources.id WHERE sources.id IS NULL)');
+        if ($date !== NULL)
+            \F3::get('db')->exec('DELETE FROM items WHERE starred=false AND datetime<:date',
                     array(':date' => $date->format('Y-m-d').' 00:00:00'));
     }
 }
